moving Storage trait to storage crate
