+++ tee-worker/omni-executor/executor-core/src/lib.rs
@@ -21,5 +21,4 @@ pub mod key_store;
 pub mod listener;
 pub mod native_call;
 pub mod primitives;
-pub mod storage;
 pub mod sync_checkpoint_repository;
