+++ tee-worker/omni-executor/parentchain/listener/src/event_handler.rs
@@ -22,7 +22,6 @@ use executor_core::event_handler::{Error, EventHandler as EventHandlerTrait};
 use executor_core::intent_executor::IntentExecutor;
 use executor_core::key_store::KeyStore;
 use executor_core::primitives::Intent;
-use executor_core::storage::Storage;
 use log::error;
 use parentchain_api_interface::{
 	omni_account::{
@@ -37,6 +36,7 @@ use parentchain_rpc_client::{SubstrateRpcClient, SubstrateRpcClientFactory};
 use primitives::{AccountId, BlockEvent, Hash, MemberAccount, TryFromSubxtType};
 use std::marker::PhantomData;
 use std::sync::Arc;
+use storage::Storage;
 use subxt::ext::scale_decode;
 use subxt::ext::scale_decode::DecodeAsFields;
 use subxt::{events::StaticEvent, Config, Metadata};
